Use DafnyConsolePrinter
