+++ Source/Dafny/DafnyOptions.cs
@@ -14,7 +14,7 @@ namespace Microsoft.Dafny {
   public class DafnyOptions : Bpl.CommandLineOptions {
 
     public static DafnyOptions Create(params string[] arguments) {
-      var result = new DafnyOptions(new Bpl.ConsolePrinter());
+      var result = new DafnyOptions(new DafnyConsolePrinter());
       result.Parse(arguments);
       return result;
     }
