+++ Source/DafnyServer/DafnyHelper.cs
@@ -12,24 +12,6 @@ using DafnyServer;
 using Bpl = Microsoft.Boogie;
 
 namespace Microsoft.Dafny {
-  // FIXME: This should not be duplicated here
-  class DafnyConsolePrinter : ConsolePrinter {
-
-    public override void ReportBplError(IToken tok, string message, bool error, TextWriter tw, string category = null) {
-      // Dafny has 0-indexed columns, but Boogie counts from 1
-      var realigned_tok = new Token(tok.line, tok.col - 1);
-      realigned_tok.kind = tok.kind;
-      realigned_tok.pos = tok.pos;
-      realigned_tok.val = tok.val;
-      realigned_tok.filename = tok.filename;
-      base.ReportBplError(realigned_tok, message, error, tw, category);
-
-      if (tok is Dafny.NestedToken) {
-        var nt = (Dafny.NestedToken)tok;
-        ReportBplError(nt.Inner, "Related location", false, tw);
-      }
-    }
-  }
 
   class DafnyHelper {
     private string fname;
